feat: GetInitCodeFromFactoryCallData, ref
