+++ pkg/smart_wallet/init_code.go
@@ -1,80 +1,13 @@
 package smart_wallet
 
 import (
-	"context"
-	"errors"
 	"fmt"
 	"log/slog"
 
-	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/shopspring/decimal"
 )
-
-var (
-	res, ok                  = entryPointABI.Errors["SenderAddressResult"]
-	senderAddressResultError = boolMust(res, ok)
-)
-
-func IsAccountDeployed(ctx context.Context, provider ethereum.ChainStateReader, swAddress common.Address) (bool, error) {
-	byteCode, err := provider.CodeAt(ctx, swAddress, nil)
-	if err != nil {
-		return false, fmt.Errorf("failed to check if smart account is already deployed: %w", err)
-	}
-
-	// assume that the smart account is deployed if it has non-zero byte code
-	return len(byteCode) != 0, nil
-}
-
-func GetAccountAddress(ctx context.Context, provider ethereum.ContractCaller, config Config, entryPointAddress, owner common.Address, index decimal.Decimal) (common.Address, error) {
-	initCode, err := GetInitCode(config, owner, index)
-	if err != nil {
-		return common.Address{}, fmt.Errorf("failed to build initCode middleware: %w", err)
-	}
-
-	// calculate the smart wallet address that will be generated by the entry point
-	// See https://github.com/eth-infinitism/account-abstraction/blob/v0.6.0/contracts/core/EntryPoint.sol#L356
-	getSenderAddressData, err := entryPointABI.Pack("getSenderAddress", initCode)
-	if err != nil {
-		return common.Address{}, fmt.Errorf("failed to pack getSenderAddress data: %w", err)
-	}
-
-	msg := ethereum.CallMsg{
-		To:   &entryPointAddress,
-		Data: getSenderAddressData,
-	}
-
-	// this call must always revert (see EntryPoint contract), so we expect an error
-	_, err = provider.CallContract(ctx, msg, nil)
-	if err == nil {
-		panic(fmt.Errorf("'getSenderAddress' call returned no error, but expected one"))
-	}
-
-	var scError rpc.DataError
-	if ok := errors.As(err, &scError); !ok {
-		return common.Address{}, fmt.Errorf("unexpected error type '%T' containing message %w)", err, err)
-	}
-	errorData := scError.ErrorData().(string)
-
-	// check if the error signature is correct
-	if id := senderAddressResultError.ID.String(); errorData[0:10] != id[0:10] {
-		return common.Address{}, fmt.Errorf("'getSenderAddress' unexpected error signature: %s", errorData[0:10])
-	}
-
-	// check if the error data has the correct length
-	if len(errorData) < 74 {
-		return common.Address{}, fmt.Errorf("'getSenderAddress' revert data expected to have lenght of 74, but got: %d", len(errorData))
-	}
-
-	swAddress := common.HexToAddress(errorData[34:])
-	if swAddress == (common.Address{}) {
-		return common.Address{}, fmt.Errorf("'getSenderAddress' returned zero address")
-	}
-
-	return swAddress, nil
-}
 
 func GetInitCode(smartWalletConfig Config, ownerAddress common.Address, index decimal.Decimal) ([]byte, error) {
 	var initCode []byte
@@ -98,7 +31,7 @@ func GetInitCode(smartWalletConfig Config, ownerAddress common.Address, index de
 	return initCode, nil
 }
 
-func GetFactoryCallData(provider *ethclient.Client, smartWalletConfig Config, ownerAddress common.Address, index decimal.Decimal) ([]byte, error) {
+func GetInitCodeFromFactoryCallData(smartWalletConfig Config, factoryCallData []byte) ([]byte, error) {
 	var initCode []byte
 	var err error
 
@@ -106,9 +39,9 @@ func GetFactoryCallData(provider *ethclient.Client, smartWalletConfig Config, ow
 	case SimpleAccountType:
 		return nil, fmt.Errorf("%w: %s", ErrSmartWalletNotSupported, typ)
 	case BiconomyType: // not tested
-		initCode, err = GetBiconomyFactoryCallData(ownerAddress, index, smartWalletConfig.ECDSAValidator)
+		initCode, err = GetBiconomyInitCodeFromFactoryCallData(smartWalletConfig.Factory, factoryCallData)
 	case KernelType:
-		initCode, err = GetKernelFactoryCallData(ownerAddress, index, smartWalletConfig.Logic, smartWalletConfig.ECDSAValidator)
+		initCode, err = GetKernelInitCodeFromFactoryCallData(smartWalletConfig.Factory, factoryCallData)
 	default:
 		return nil, fmt.Errorf("unknown smart wallet type: %s", typ)
 	}
@@ -120,7 +53,7 @@ func GetFactoryCallData(provider *ethclient.Client, smartWalletConfig Config, ow
 	return initCode, nil
 }
 
-// getKernelInitCode the init code for a Zerodev Kernel smart account.
+// GetKernelInitCode the init code for a Zerodev Kernel smart account.
 // The init code deploys a smart account if it is not already deployed.
 func GetKernelInitCode(owner common.Address, index decimal.Decimal, factory, accountLogic, ecdsaValidator common.Address) ([]byte, error) {
 	callData, err := GetKernelFactoryCallData(owner, index, accountLogic, ecdsaValidator)
@@ -128,38 +61,24 @@ func GetKernelInitCode(owner common.Address, index decimal.Decimal, factory, acc
 		return nil, fmt.Errorf("failed to get init data: %w", err)
 	}
 
+	return GetKernelInitCodeFromFactoryCallData(factory, callData)
+}
+
+// GetKernelInitCodeFromFactoryCallData returns the init code for a Zerodev Kernel smart account,
+// given the Kernel Factory calldata. The init code deploys a smart account if it is not already deployed.
+func GetKernelInitCodeFromFactoryCallData(factory common.Address, factoryCallData []byte) ([]byte, error) {
 	// Pack factory address and deployment data for `CreateSender` in EntryPoint
 	// see https://github.com/eth-infinitism/account-abstraction/blob/v0.6.0/contracts/core/SenderCreator.sol#L15
-	initCode := make([]byte, len(factory)+len(callData))
+	initCode := make([]byte, len(factory)+len(factoryCallData))
 	copy(initCode, factory.Bytes())
-	copy(initCode[len(factory):], callData)
+	copy(initCode[len(factory):], factoryCallData)
 
 	slog.Debug("built initCode", "initCode", hexutil.Encode(initCode))
 
 	return initCode, nil
 }
 
-// getKernelInitData returns the calldata needed call the factory
-// to deploy a Zerodev Kernel smart account.
-func GetKernelFactoryCallData(owner common.Address, index decimal.Decimal, accountLogic, ecdsaValidator common.Address) ([]byte, error) {
-	// Initialize Kernel Smart Account with default validation module and its calldata
-	// see https://github.com/zerodevapp/kernel/blob/807b75a4da6fea6311a3573bc8b8964a34074d94/src/abstract/KernelStorage.sol#L35
-	initData, err := kernelInitABI.Pack("initialize", ecdsaValidator, owner.Bytes())
-	if err != nil {
-		return nil, fmt.Errorf("failed to pack init data: %w", err)
-	}
-
-	// Deploy Kernel Smart Account by calling `factory.createAccount`
-	// see https://github.com/zerodevapp/kernel/blob/807b75a4da6fea6311a3573bc8b8964a34074d94/src/factory/KernelFactory.sol#L25
-	callData, err := kernelDeployWalletABI.Pack("createAccount", accountLogic, initData, index.BigInt())
-	if err != nil {
-		return nil, fmt.Errorf("failed to pack createAccount data: %w", err)
-	}
-
-	return callData, nil
-}
-
-// getBiconomyInitCode returns the init code for a Biconomy smart account.
+// GetBiconomyInitCode returns the init code for a Biconomy smart account.
 // The init code deploys a smart account if it is not already deployed.
 // NOTE: this was NOT tested. Use at your own risk or wait for the package to be updated.
 func GetBiconomyInitCode(owner common.Address, index decimal.Decimal, factory, ecdsaValidator common.Address) ([]byte, error) {
@@ -168,33 +87,19 @@ func GetBiconomyInitCode(owner common.Address, index decimal.Decimal, factory, e
 		return nil, fmt.Errorf("failed to get init data: %w", err)
 	}
 
+	return GetBiconomyInitCodeFromFactoryCallData(factory, callData)
+}
+
+// GetBiconomyInitCodeFromFactoryCallData returns the init code for a Biconomy smart account,
+// given the Biconomy Factory calldata. The init code deploys a smart account if it is not already deployed.
+func GetBiconomyInitCodeFromFactoryCallData(factory common.Address, factoryCallData []byte) ([]byte, error) {
 	// Pack factory address and deployment data for `CreateSender` in EntryPoint
 	// see https://github.com/eth-infinitism/account-abstraction/blob/v0.6.0/contracts/core/SenderCreator.sol#L15
-	initCode := make([]byte, len(factory)+len(callData))
+	initCode := make([]byte, len(factory)+len(factoryCallData))
 	copy(initCode, factory.Bytes())
-	copy(initCode[len(factory):], callData)
+	copy(initCode[len(factory):], factoryCallData)
 
 	slog.Debug("built initCode", "initCode", hexutil.Encode(initCode))
 
 	return initCode, nil
-}
-
-// getKernelInitData returns the calldata needed call the factory
-// to deploy a Biconomy smart account.
-func GetBiconomyFactoryCallData(owner common.Address, index decimal.Decimal, ecdsaValidator common.Address) ([]byte, error) {
-	// Initialize SCW validation module with owner address
-	// see https://github.com/bcnmy/scw-contracts/blob/v2-deployments/contracts/smart-account/modules/EcdsaOwnershipRegistryModule.sol#L43
-	ecdsaOwnershipInitData, err := biconomyInitABI.Pack("initForSmartAccount", owner.Bytes())
-	if err != nil {
-		return nil, fmt.Errorf("failed to pack init data: %w", err)
-	}
-
-	// Deploy Biconomy SCW by calling `factory.createAccount`
-	// see https://github.com/bcnmy/scw-contracts/blob/v2-deployments/contracts/smart-account/factory/SmartAccountFactory.sol#L112
-	callData, err := biconomyDeployWalletABI.Pack("createAccount", ecdsaValidator, ecdsaOwnershipInitData, index.BigInt())
-	if err != nil {
-		return nil, fmt.Errorf("failed to pack createAccount data: %w", err)
-	}
-
-	return callData, nil
 }
