+++ pkg/userop/testing/local_blockchain/node.go
@@ -4,16 +4,13 @@ import (
 	"context"
 	"fmt"
 	"log/slog"
-	"math/big"
 	"net/url"
 	"os"
 	"sync"
 	"testing"
-	"time"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/shopspring/decimal"
@@ -339,59 +336,4 @@ func buildClient(t *testing.T, rpcURL, bundlerURL url.URL, addresses Contracts) 
 	client, err := userop.NewClient(config)
 	require.NoError(t, err)
 	return client
-}
-
-func sendNative(ctx context.Context, node *EthNode, from, to Account, fundAmount *big.Int) error {
-	chainID, err := node.Client.ChainID(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to get chain ID: %w", err)
-	}
-
-	nonce, err := node.Client.PendingNonceAt(ctx, from.Address)
-	if err != nil {
-		return fmt.Errorf("failed to get nonce: %w", err)
-	}
-
-	gasLimit := uint64(21000)
-	gasPrice, err := node.Client.SuggestGasPrice(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to suggest gas price: %w", err)
-	}
-
-	tx := types.NewTransaction(nonce, to.Address, fundAmount, gasLimit, gasPrice, nil)
-	signedTx, err := types.SignTx(tx, types.NewEIP155Signer(chainID), from.PrivateKey)
-	if err != nil {
-		return fmt.Errorf("failed to sign transaction: %w", err)
-	}
-
-	err = node.Client.SendTransaction(ctx, signedTx)
-	if err != nil {
-		return fmt.Errorf("failed to send transaction: %w", err)
-	}
-
-	_, err = waitMined(ctx, node, signedTx)
-	if err != nil {
-		return fmt.Errorf("failed to wait for transaction to be mined: %w", err)
-	}
-
-	return nil
-}
-
-func waitMined(ctx context.Context, node *EthNode, tx *types.Transaction) (*types.Receipt, error) {
-	queryTicker := time.NewTicker(1 * time.Second)
-	defer queryTicker.Stop()
-
-	for {
-		receipt, err := node.Client.TransactionReceipt(ctx, tx.Hash())
-		if err == nil {
-			return receipt, nil
-		}
-
-		// Wait for the next round.
-		select {
-		case <-ctx.Done():
-			return nil, ctx.Err()
-		case <-queryTicker.C:
-		}
-	}
 }
